pyproject.toml
[tool.poetry]
name = "person-api"
version = "0.1.0"
description = ""
authors = ["Vitor <lagevitor.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
pyyaml = "^6.0"
sqlalchemy = "^1.4.31"
pydantic = "^1.8.2"
psycopg2 = "^2.9.1" 

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
